feat: add vuex for global state management
